int to float price

+++ stocks_table.py
@@ -1,6 +1,6 @@
 import os
 
-from sqlalchemy import create_engine, Column, Integer, String
+from sqlalchemy import create_engine, Column, Integer, String,Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from dotenv import load_dotenv
@@ -18,23 +18,25 @@ DATABASE_URL = f"postgresql://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
 engine = create_engine(DATABASE_URL)
 Base=declarative_base()
 
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
+
 class Stock(Base):
     __tablename__ = 'stocks'
-    id = Column(Integer, primary_key=True)
-    name = Column(String)
-    symbol = Column(String,index=True)
-    sector = Column(String,index=True)
-    open = Column(Integer)
-    high = Column(Integer)
-    low = Column(Integer)
-    current_price = Column(Integer)
-    change = Column(Integer)
-    volume = Column(Integer)
-    beta = Column(Integer)
+    id            = Column(Integer, primary_key=True)
+    name          = Column(String)
+    symbol        = Column(String, index=True)
+    sector        = Column(String, index=True)
+    open          = Column(Float)        # ✅ was Integer
+    high          = Column(Float)        # ✅
+    low           = Column(Float)        # ✅
+    current_price = Column(Float)        # ✅
+    change        = Column(Float)        # ✅
+    volume        = Column(Integer)      # ✅ keep int — volume is always whole number
+    beta          = Column(Float)        # ✅
     shariah_status = Column(String)
-    risk_level = Column(String,index=True)
-    timestamp = Column(String)
-
+    risk_level    = Column(String, index=True)
+    timestamp     = Column(String)
 
 Base.metadata.create_all(engine)
